requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
lxml==4.9.4
numpy==1.26.2
oauthlib==3.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==2.1.0
